+++ backend/src/config/redis.ts
@@ -10,15 +10,19 @@ export const cacheKeySubjectData = "subject_data";
 export const cacheKeyTeamData = "teams_data";
 export const cacheExpiration = 3600;
 
-const redisHost = process.env.NODE_ENV === "DEVELOPMENT" ? "localhost" : "redis";
+const redisHost =
+  process.env.NODE_ENV === "DEVELOPMENT" ? "localhost" : "redis";
 const redisPort = process.env.REDIS_PORT || "6379";
 const redisUrl = `redis://${redisHost}:${redisPort}`;
 
 export const redisClient = createClient({
   url: redisUrl,
 });
-redisClient.on("error", (err: unknown) => (err instanceof Error) ? logger.error("Redis error:", err) : logger.error("Unknown Redis error!"));
-
+redisClient.on("error", (err: unknown) =>
+  err instanceof Error
+    ? logger.error("Redis error:", err)
+    : logger.error("Unknown Redis error!")
+);
 
 export const connectRedis = async (): Promise<void> => {
   try {
@@ -26,8 +30,7 @@ export const connectRedis = async (): Promise<void> => {
       await redisClient.connect();
       logger.success("Connected to Redis");
     }
-  }
-  catch (err: unknown) {
+  } catch (err: unknown) {
     if (err instanceof Error) {
       logger.error("Error connecting to Redis:", err);
       throw err;