+++ backend/src/controllers/lessonController.ts
@@ -9,7 +9,7 @@ export const getLessonData = asyncHandler(async (req, res, next) => {
   } catch (error) {
     next(error);
   }
-})
+});
 export const setLessonData = asyncHandler(async (req, res, next) => {
   const setEventTypesSchema = z.object({
     lessons: z.array(
@@ -20,12 +20,17 @@ export const setLessonData = asyncHandler(async (req, res, next) => {
         subjectId: z.coerce.number(),
         room: z.string(),
         startTime: z.coerce.number(),
-        endTime: z.coerce.number()
+        endTime: z.coerce.number(),
       })
-    )
-  })
-  let parseResult = setEventTypesSchema.safeParse(req.body);
-  if (! parseResult.success || parseResult.data.lessons.map(l => [0, 1, 2, 3, 4].includes(l.weekDay)).includes(false)) {
+    ),
+  });
+  const parseResult = setEventTypesSchema.safeParse(req.body);
+  if (
+    !parseResult.success ||
+    parseResult.data.lessons
+      .map(l => [0, 1, 2, 3, 4].includes(l.weekDay))
+      .includes(false)
+  ) {
     res.status(400).json({
       error: "Invalid request format",
       expectedFormat: {
@@ -37,21 +42,29 @@ export const setLessonData = asyncHandler(async (req, res, next) => {
               type: "object",
               properties: {
                 lessonNumber: { type: "number" },
-                weekDay: { enum: [ 0, 1, 2, 3, 4, "0", "1", "2", "3", "4"] },
+                weekDay: { enum: [0, 1, 2, 3, 4, "0", "1", "2", "3", "4"] },
                 teamId: { type: "number" },
                 subjectId: { type: "number" },
                 room: { type: "string" },
                 startTime: { type: "number" },
-                endTime: { type: "number" }
+                endTime: { type: "number" },
               },
-              required: ["lessonNumber", "weekDay", "teamId", "subjectId", "room", "startTime", "endTime"]
-            }
-          }
+              required: [
+                "lessonNumber",
+                "weekDay",
+                "teamId",
+                "subjectId",
+                "room",
+                "startTime",
+                "endTime",
+              ],
+            },
+          },
         },
-        required: ["lessons"]
-      }
+        required: ["lessons"],
+      },
     });
-    return
+    return;
   }
   try {
     await lessonService.setLessonData(parseResult.data.lessons);
@@ -59,9 +72,9 @@ export const setLessonData = asyncHandler(async (req, res, next) => {
   } catch (error) {
     next(error);
   }
-})
+});
 
 export default {
   getLessonData,
-  setLessonData
+  setLessonData,
 };