+++ backend/src/services/lessonService.ts
@@ -1,38 +1,49 @@
-const fs = require("fs").promises;
 import { RequestError } from "../@types/requestError";
 import { cacheKeyLessonData, redisClient } from "../config/redis";
 import prisma from "../config/prisma";
 import logger from "../utils/logger";
-import { isValidweekDay, BigIntreplacer, updateCacheData } from "../utils/validateFunctions";
-
+import {
+  isValidweekDay,
+  BigIntreplacer,
+  updateCacheData,
+} from "../utils/validateFunctions";
 
 const lessonService = {
-  async setLessonData(lessons: { lessonNumber: number, weekDay:number, teamId: number, subjectId: number, room: string, startTime: number, endTime: number}[]) {
-    for (let lesson of lessons) {
+  async setLessonData(
+    lessons: {
+      lessonNumber: number;
+      weekDay: number;
+      teamId: number;
+      subjectId: number;
+      room: string;
+      startTime: number;
+      endTime: number;
+    }[]
+  ) {
+    for (const lesson of lessons) {
       isValidweekDay(lesson.weekDay);
     }
     await prisma.$executeRaw`TRUNCATE TABLE "lesson" RESTART IDENTITY;`;
-    for (let lesson of lessons) {
+    for (const lesson of lessons) {
       try {
-          await prisma.lesson.create({
-            data: {
-              lessonNumber: lesson.lessonNumber,
-              weekDay: lesson.weekDay as 0 | 1 | 2 | 3 | 4,
-              teamId: lesson.teamId,
-              subjectId: lesson.subjectId,
-              room: lesson.room,
-              startTime: lesson.startTime,
-              endTime: lesson.endTime
-            }
-          })
-        }
-      catch {
-        let err: RequestError = {
+        await prisma.lesson.create({
+          data: {
+            lessonNumber: lesson.lessonNumber,
+            weekDay: lesson.weekDay as 0 | 1 | 2 | 3 | 4,
+            teamId: lesson.teamId,
+            subjectId: lesson.subjectId,
+            room: lesson.room,
+            startTime: lesson.startTime,
+            endTime: lesson.endTime,
+          },
+        });
+      } catch {
+        const err: RequestError = {
           name: "Bad Request",
           status: 400,
           message: "Invalid data format",
           expected: true,
-        }
+        };
         throw err;
       }
     }
@@ -47,7 +58,7 @@ const lessonService = {
   },
   async getLessonData() {
     const cachedLessonData = await redisClient.get("lesson_data");
-    
+
     if (cachedLessonData) {
       try {
         return JSON.parse(cachedLessonData);
@@ -67,7 +78,7 @@ const lessonService = {
     }
 
     return JSON.stringify(lessonData, BigIntreplacer);
-  }
-}
+  },
+};
 
 export default lessonService;