+++ backend/src/services/teamService.ts
@@ -29,63 +29,63 @@ const teamService = {
 
     return data;
   },
-  async setTeamsData(teams: { teamId: number | "", name: string }[]) {
-    let existingTeams = await prisma.team.findMany();
-    await Promise.all(existingTeams.map(async (team: {teamId: number}) => {
-      if (!teams.some((t) => t.teamId === team.teamId)) {
-        // delete homework which were linked to team
-        await prisma.homework10d.deleteMany({
-          where: { teamId: team.teamId }
-        });
-        // delete events which were linked to team
-        await prisma.event.deleteMany({
-          where: { teamId: team.teamId }
-        });
-        // delete lessons which were linked to team
-        await prisma.lesson.deleteMany({
-          where: { teamId: team.teamId }
-        });
-        // delete team
-        await prisma.team.delete({
-          where: { teamId: team.teamId }
-        });
-      }
-    }));
+  async setTeamsData(teams: { teamId: number | ""; name: string }[]) {
+    const existingTeams = await prisma.team.findMany();
+    await Promise.all(
+      existingTeams.map(async (team: { teamId: number }) => {
+        if (!teams.some(t => t.teamId === team.teamId)) {
+          // delete homework which were linked to team
+          await prisma.homework10d.deleteMany({
+            where: { teamId: team.teamId },
+          });
+          // delete events which were linked to team
+          await prisma.event.deleteMany({
+            where: { teamId: team.teamId },
+          });
+          // delete lessons which were linked to team
+          await prisma.lesson.deleteMany({
+            where: { teamId: team.teamId },
+          });
+          // delete team
+          await prisma.team.delete({
+            where: { teamId: team.teamId },
+          });
+        }
+      })
+    );
 
-    for (let team of teams) {
+    for (const team of teams) {
       if (team.name.trim() == "") {
-        let err: RequestError = {
+        const err: RequestError = {
           name: "Bad Request",
           status: 400,
           message: "Invalid data (Team name cannot be empty)",
           expected: true,
-        }
+        };
         throw err;
       }
       try {
         if (team.teamId == "") {
           await prisma.team.create({
             data: {
-              name: team.name
-            }
-          })
-        }
-        else {
+              name: team.name,
+            },
+          });
+        } else {
           await prisma.team.update({
             where: { teamId: team.teamId },
             data: {
-              name: team.name
+              name: team.name,
             },
           });
         }
-      }
-      catch {
-        let err: RequestError = {
+      } catch {
+        const err: RequestError = {
           name: "Bad Request",
           status: 400,
           message: "Invalid data format",
           expected: true,
-        }
+        };
         throw err;
       }
     }
@@ -100,80 +100,81 @@ const teamService = {
     }
   },
   async getJoinedTeamsData(session: Session & Partial<SessionData>) {
-    let accountId
-    if (!(session.account)) {
-      let err: RequestError = {
+    let accountId;
+    if (!session.account) {
+      const err: RequestError = {
         name: "Unauthorized",
         status: 401,
         message: "User not logged in",
         expected: true,
-      }
-      throw err;
-    }
-    else {
-      accountId = session.account.accountId;
-    }
-
-    const data = await prisma.joinedTeams.findMany({
-      where: { accountId: accountId }
-    });
-
-    let teams = []
-
-    for (let entry of data) {
-      teams.push(entry.teamId);
-    };
-
-    return teams;
-  },
-  async setJoinedTeamsData(teams: number[], session: Session & Partial<SessionData>) {
-    let accountId
-    if (!(session.account)) {
-      let err: RequestError = {
-        name: "Unauthorized",
-        status: 401,
-        message: "User not logged in",
-        expected: true,
-      }
+      };
       throw err;
     } else {
       accountId = session.account.accountId;
     }
 
-    if (! Array.isArray(teams)) {
-      let err: RequestError = {
+    const data = await prisma.joinedTeams.findMany({
+      where: { accountId: accountId },
+    });
+
+    const teams = [];
+
+    for (const entry of data) {
+      teams.push(entry.teamId);
+    }
+
+    return teams;
+  },
+  async setJoinedTeamsData(
+    teams: number[],
+    session: Session & Partial<SessionData>
+  ) {
+    let accountId;
+    if (!session.account) {
+      const err: RequestError = {
+        name: "Unauthorized",
+        status: 401,
+        message: "User not logged in",
+        expected: true,
+      };
+      throw err;
+    } else {
+      accountId = session.account.accountId;
+    }
+
+    if (!Array.isArray(teams)) {
+      const err: RequestError = {
         name: "Bad Request",
         status: 400,
         message: "Invalid data format",
         expected: true,
-      }
+      };
       throw err;
     }
 
-    await prisma.joinedTeams.delete({
-      where: { accountId: accountId }
+    await prisma.joinedTeams.deleteMany({
+      where: { accountId: accountId },
     });
 
-    for (let teamId of teams) {
+    for (const teamId of teams) {
       try {
         await prisma.joinedTeams.create({
           data: {
             teamId: teamId,
-            accountId: accountId
-          }
-        })
-      }
-      catch {
-        let err: RequestError = {
+            accountId: accountId,
+          },
+        });
+      } catch {
+        const err: RequestError = {
           name: "Bad Request",
           status: 400,
           message: "Invalid data format",
           expected: true,
-        }
+        };
         throw err;
       }
     }
   },
-}
+};
 
 export default teamService;